Feature/ksa role permission authurization

+++ app/Http/Controllers/Api/v1/AssetTypeApiController.php
@@ -1,11 +1,15 @@
 <?php
 
 namespace App\Http\Controllers\Api\v1;
-use App\Models\AssetType;
-use Illuminate\Http\Request;
+
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AssetTypeRequest;
+use App\Models\AssetType;
+use App\Services\FilterQueryService;
 use App\Traits\ResponserTrait;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class AssetTypeApiController extends Controller
 {
@@ -13,6 +17,8 @@ class AssetTypeApiController extends Controller
 
     public function index()
     {
+        abort_if(Gate::denies('asset_type_access'), $this->respondPermissionDenied());
+
         $asset_types = AssetType::paginate(10);
 
         return $this->respondCollectionWithPagination('success', $asset_types);
@@ -20,6 +26,8 @@ class AssetTypeApiController extends Controller
 
     public function store(AssetTypeRequest $request)
     {
+        abort_if(Gate::denies('asset_type_create'), $this->respondPermissionDenied());
+
         AssetType::create([
             'asset_type' => $request->asset_type,
         ]);
@@ -27,7 +35,7 @@ class AssetTypeApiController extends Controller
         return $this->respondCreateMessageOnly('success');
     }
 
-     /**
+    /**
      * Display the specified resource.
      *
      * @param  \App\Models\AssetType  $assetType
@@ -40,6 +48,8 @@ class AssetTypeApiController extends Controller
 
     public function update(Request $request, AssetType $assetType)
     {
+        abort_if(Gate::denies('asset_type_update'), $this->respondPermissionDenied());
+
         $assetType->update([
             'asset_type' => $request->asset_type,
             'asset_type' => $request->asset_type,
@@ -48,7 +58,7 @@ class AssetTypeApiController extends Controller
         return $this->respondCreateMessageOnly('success');
     }
 
-      /**
+    /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Models\AssetType  $assetType
@@ -56,8 +66,21 @@ class AssetTypeApiController extends Controller
      */
     public function destroy(AssetType $assetType)
     {
+        abort_if(Gate::denies('asset_type_delete'), $this->respondPermissionDenied());
+
         $assetType->delete();
 
         return $this->respondCreateMessageOnly('success');
+    }
+
+    public function query(Request $request)
+    {
+        abort_if(Gate::denies('asset_type_query'), $this->respondPermissionDenied());
+
+        $asset_types = DB::table('asset_types');
+
+        $data = FilterQueryService::FilterQuery($request, $asset_types);
+
+        return $this->respondCollectionWithPagination('success', $data);
     }
 }
